Move config modifying utils to new files as they're hella long
